+++ packages/contracts/src/framework/plugin/repo/PluginRepoFactory.sol
@@ -7,7 +7,7 @@ import {ERC165} from "@openzeppelin/contracts/utils/introspection/ERC165.sol";
 import {PermissionLib} from "@aragon/osx-commons-contracts/src/permission/PermissionLib.sol";
 import {IProtocolVersion} from "@aragon/osx-commons-contracts/src/utils/versioning/IProtocolVersion.sol";
 import {ProtocolVersion} from "@aragon/osx-commons-contracts/src/utils/versioning/ProtocolVersion.sol";
-import {createERC1967Proxy} from "@aragon/osx-commons-contracts/src/utils/deployment/Proxy.sol";
+import {ProxyLib} from "@aragon/osx-commons-contracts/src/utils/deployment/ProxyLib.sol";
 import {PluginRepoRegistry} from "./PluginRepoRegistry.sol";
 import {PluginRepo} from "./PluginRepo.sol";
 
@@ -16,6 +16,8 @@ import {PluginRepo} from "./PluginRepo.sol";
 /// @notice This contract creates `PluginRepo` proxies and registers them on a `PluginRepoRegistry` contract.
 /// @custom:security-contact sirt@aragon.org
 contract PluginRepoFactory is ERC165, ProtocolVersion {
+    using ProxyLib for address;
+
     /// @notice The Aragon plugin registry contract.
     PluginRepoRegistry public pluginRepoRegistry;
 
@@ -131,10 +133,7 @@ contract PluginRepoFactory is ERC165, ProtocolVersion {
         address _initialOwner
     ) internal returns (PluginRepo pluginRepo) {
         pluginRepo = PluginRepo(
-            createERC1967Proxy(
-                pluginRepoBase,
-                abi.encodeCall(PluginRepo.initialize, (_initialOwner))
-            )
+            pluginRepoBase.deployUUPSProxy(abi.encodeCall(PluginRepo.initialize, (_initialOwner)))
         );
 
         pluginRepoRegistry.registerPluginRepo(_subdomain, address(pluginRepo));