+++ packages/contracts/src/test/Migration.sol
@@ -56,6 +56,6 @@ import {GovernanceWrappedERC20 as GovernanceWrappedERC20_v1_0_0} from "@aragon/o
 import {GovernanceWrappedERC20 as GovernanceWrappedERC20_v1_3_0} from "@aragon/osx-v1.3.0/token/ERC20/governance/GovernanceWrappedERC20.sol";
 
 // Integration Testing
-import {CloneFactory} from "@aragon/osx-commons-contracts/src/utils/deployment/CloneFactory.sol";
+import {ProxyFactory} from "@aragon/osx-commons-contracts/src/utils/deployment/ProxyFactory.sol";
 
 /* solhint-enable no-unused-import */