+++ packages/contracts/src/test/plugin/Cloneable/PluginCloneableSetupMock.sol
@@ -11,18 +11,14 @@ import {mockPermissions, mockHelpers, mockPluginProxy} from "../PluginMockData.s
 import {PluginCloneableV1Mock, PluginCloneableV1MockBad, PluginCloneableV2Mock} from "./PluginCloneableMock.sol";
 
 contract PluginCloneableSetupV1Mock is PluginSetup {
-    address internal pluginBase;
-
-    constructor() {
-        pluginBase = address(new PluginCloneableV1Mock());
-    }
+    constructor(address implementation) PluginSetup(implementation) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
     ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
-        plugin = mockPluginProxy(pluginBase, _dao);
+        plugin = mockPluginProxy(implementation(), _dao);
         preparedSetupData.helpers = mockHelpers(1);
         preparedSetupData.permissions = mockPermissions(5, 6, PermissionLib.Operation.Grant);
     }
@@ -35,30 +31,40 @@ contract PluginCloneableSetupV1Mock is PluginSetup {
         (_dao, _payload);
         permissions = mockPermissions(5, 6, PermissionLib.Operation.Revoke);
     }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view virtual override returns (address) {
-        return address(pluginBase);
-    }
 }
 
-contract PluginCloneableSetupV1MockBad is PluginCloneableSetupV1Mock {
-    constructor() {
-        pluginBase = address(new PluginCloneableV1MockBad());
-    }
-}
-
-contract PluginCloneableSetupV2Mock is PluginCloneableSetupV1Mock {
-    constructor() {
-        pluginBase = address(new PluginCloneableV2Mock());
-    }
+contract PluginCloneableSetupV1MockBad is PluginSetup {
+    constructor(address implementation) PluginSetup(implementation) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
     ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
-        plugin = mockPluginProxy(pluginBase, _dao);
+        plugin = mockPluginProxy(implementation(), _dao);
+        preparedSetupData.helpers = mockHelpers(1);
+        preparedSetupData.permissions = mockPermissions(5, 6, PermissionLib.Operation.Grant);
+    }
+
+    /// @inheritdoc IPluginSetup
+    function prepareUninstallation(
+        address _dao,
+        SetupPayload calldata _payload
+    ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+        (_dao, _payload);
+        permissions = mockPermissions(5, 6, PermissionLib.Operation.Revoke);
+    }
+}
+
+contract PluginCloneableSetupV2Mock is PluginCloneableSetupV1Mock {
+    constructor(address implementation) PluginCloneableSetupV1Mock(implementation) {}
+
+    /// @inheritdoc IPluginSetup
+    function prepareInstallation(
+        address _dao,
+        bytes memory
+    ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+        plugin = mockPluginProxy(implementation(), _dao);
         preparedSetupData.helpers = mockHelpers(1);
         preparedSetupData.permissions = mockPermissions(5, 7, PermissionLib.Operation.Grant);
     }
@@ -70,10 +76,5 @@ contract PluginCloneableSetupV2Mock is PluginCloneableSetupV1Mock {
     ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
         permissions = mockPermissions(5, 7, PermissionLib.Operation.Revoke);
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view virtual override returns (address) {
-        return address(pluginBase);
     }
 }